Fix scheduling algorithm
The issue was that in some situations, the scheduler would not flush
stream B if there was a stream A with higher priority even if stream
A had no data to send. This would not be an issue if streams were
constantly sending data, but if all streams stopped sending, frames
might be stuck in the scheduler.

The fix to this is to traverse the children until a stream that needs
to send data is found.

Also fixes a possibly broken test. After the changes to the scheduler,
a frame that wasn't flushed before is now flushed.